Use prefix rather than RGSM to find DRAGEN inputs
RGSM was used in the original implementation to find DRAGEN input files
however this only works when DRAGEN is configured to run with RGSM (or
tumour RGSM) and output prefix set as the same value. This is often the
case but not always. Updating to use output prefix rather than RGSM.